Move ideal gas model into twine-components and do some other stuff
