+++ twine-components/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod example;
+pub mod fluid;
 pub mod interpolation;
 pub mod solver;
