+++ twine-core/Cargo.toml
@@ -11,4 +11,5 @@ keywords = ["twine", "framework", "functional", "composable", "modeling"]
 
 [dependencies]
 petgraph = "0.7.1"
+thiserror = "2.0.12"
 uom = "0.36.0"
